refactor(lfs): share verified discovery with fetch and pull
